refactor: remove MathfUtilities, introduce Limits2 type
