+++ CommunityProject/Assets/_Game/Scripts/Runtime/Infrastructure/Limits2.cs
@@ -1,0 +1,14 @@
+namespace BoundfoxStudios.CommunityProject.Infrastructure
+{
+	public struct Limits2
+	{
+		public float Minimum { get; set; }
+		public float Maximum { get; set; }
+
+		public Limits2(float minimum, float maximum)
+		{
+			Minimum = minimum;
+			Maximum = maximum;
+		}
+	}
+}
