+++ CommunityProject/Assets/_Game/Scripts/Runtime/Infrastructure/Limits2.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a23f7f8f2e1943e6b56ae0ef5a5bb5f1
+timeCreated: 1675150674
