--- CommunityProject/Assets/_Game/Scripts/Runtime/Utilities.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 15b3604d93b845abbb5d7545c667094d
-timeCreated: 1674827406
