--- CommunityProject/Assets/_Game/Scripts/Runtime/Utilities/MathfUtilities.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 59062c7211a34b7b87669b7bdf3ff027
-timeCreated: 1674827371
